requirements.txt
Babel==2.9.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cycler==0.10.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
kiwisolver==1.3.1
matplotlib==3.3.3
mysql-connector-python==8.0.22
numpy==1.19.5
openpyxl==3.0.5
paramiko==2.7.2
Pillow==8.1.0
pkg-resources==0.0.0
protobuf==3.14.0
pycparser==2.20
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
reportlab==3.5.59
requests==2.25.1
six==1.15.0
sshtunnel==0.4.0
tkcalendar==1.6.1
urllib3==1.26.2
